AJAX dataTable - category - code refactoring
