+++ app/Services/Implement/CategoryServiceImpl.php
@@ -17,38 +17,14 @@ class CategoryServiceImpl extends BaseServiceImpl implements CategoryService
         return CategoryRepository::class;
     }
 
-    public function ajaxIndex($request)
-    {
-        return $this->makeRepo()->ajaxIndex($request);
-    }
-
-    public function ajaxEdit($id)
-    {
-        /*
-        try {
-            return $this->makeRepo()->ajaxEdit($id);
-        } catch (\Exception $e) {
-            return null;
-        }
-        */
-        return $this->makeRepo()->ajaxEdit($id);
-    }
-
-    public function ajaxUpdate($request, $id)
-    {
-        return $this->makeRepo()->ajaxUpdate($request, $id);
-    }
-
     public function ajaxStore($request)
     {
         return $this->makeRepo()->ajaxStore($request);
     }
 
-    public function ajaxDelete($id)
-    {
-        return $this->makeRepo()->ajaxDelete($id);
-    }
-
+    /**
+     * Make Model Class.
+     */
     protected function makeRepo()
     {
         return app()->make($this->getModelRepository());
