+++ resources/views/backend/admin/categories/partials/btn_action.blade.php
@@ -1,0 +1,6 @@
+<a href="javascript:void(0)" data-toggle="tooltip" data-id="{{ $id }}" data-original-title="Edit" id="edit_category" class="btn btn-info">
+    <i class="fa fa-edit" aria-hidden="true"></i>
+</a>
+<a href="javascript:void(0);" id="delete_category" data-toggle="tooltip" data-original-title="Delete" data-id="{{ $id }}" class="btn btn-danger">
+    <i class="fa fa-trash" aria-hidden="true"></i>
+</a>
