+++ resources/views/backend/admin/categories/partials/btn_trash.blade.php
@@ -1,0 +1,6 @@
+<a href="javascript:void(0)" data-toggle="tooltip" data-id="{{ $id }}" data-original-title="Restore" id="restore_category" class="btn btn-warning">
+    <i class="fa fa-undo" aria-hidden="true"></i>
+</a>
+<a href="javascript:void(0);" data-toggle="tooltip" data-id="{{ $id }}" data-original-title="Delete Permanently" id="permanent_delete_category" class="btn btn-danger">
+    <i class="fa fa-trash" aria-hidden="true"></i>
+</a>
