+++ resources/views/backend/admin/categories/partials/modal.blade.php
@@ -1,0 +1,55 @@
+<!-- Create/Update Modal -->
+<div class="modal fade" id="categoryModal" aria-hidden="true">
+    <div class="modal-dialog modal-md">
+        <div class="modal-content">
+            <div class="modal-header">
+                <h4 class="modal-title" id="categoryModalTitle"></h4>
+                <button type="button" class="close" data-dismiss="modal" aria-label="Close">
+                    <span aria-hidden="true">&times;</span>
+                </button>
+            </div>
+            <form id="categoryForm" name="categoryForm" class="form-horizontal">
+                <div class="modal-body">
+                    <div class="alert alert-warning alert-dismissible fade show">
+                        <button type="button" class="close" data-hide="alert" aria-label="Close">
+                            <span aria-hidden="true">&times;</span>
+                        </button>
+                        <div id="alert_msg" role="alert"></div>
+                    </div>
+                    <input type="hidden" name="category_id" id="category_id" />
+                    <div class="form-group">
+                        <label for="name" class="col-sm-2 control-label">Name</label>
+                        <div class="col-sm-12">
+                            <input type="text" class="form-control" id="name" name="name" placeholder="Enter Name" value="" maxlength="50" data-rule-required="true" />
+                        </div>
+                    </div>
+                </div>
+                <div class="modal-footer">
+                    <button type="button" class="btn btn-secondary" data-dismiss="modal">Close</button>
+                    <button type="submit" type="button" class="btn btn-primary" id="btn_save" value="create">Save changes</button>
+                </div>
+            </form>
+        </div>
+    </div>
+</div>
+
+<!-- Confirm Modal -->
+<div class="modal fade" id="confirmModal" tabindex="-1" role="dialog" aria-hidden="true">
+    <div class="modal-dialog" role="document">
+        <div class="modal-content">
+            <div class="modal-header">
+                <h5 class="modal-title">Confirmation</h5>
+                <button type="button" class="close" data-dismiss="modal" aria-label="Close">
+                    <span aria-hidden="true">&times;</span>
+                </button>
+            </div>
+            <div class="modal-body">
+                <h4 style="margin: 0;" id="confirmModalTitle">Are you sure you want to remove this data?</h4>
+            </div>
+            <div class="modal-footer">
+                <button type="button" class="btn btn-default" data-dismiss="modal">Cancel</button>
+                <button type="button" name="btn_ok" id="btn_ok" class="btn btn-danger">OK</button>
+            </div>
+        </div>
+    </div>
+</div>
