--- resources/views/categories/action.blade.php
@@ -1,10 +1,0 @@
-
-<a href="javascript:void(0)" data-toggle="tooltip"  data-id="{{ $id }}" 
-data-name="{{ $name }}"
-data-original-title="Edit" id="edit_category" class=" btn btn-info">
-<i class="fa fa-edit" aria-hidden="true"></i>
-</a>
-<a href="javascript:void(0);" id="delete_category" data-toggle="tooltip" 
-data-original-title="Delete" data-id="{{ $id }}" class=" btn btn-danger">
-<i class="fa fa-trash" aria-hidden="true"></i>
-</a>
