--- resources/views/categories/ajax_disable.js
@@ -1,11 +1,0 @@
-$("#productForm #name").keydown(function() {
-    var str = $("#name").val();
-     if(str != '' && $.trim(str).length >= 2){
-       $('#btn-save').prop("disabled", false);
-       }
-     else{
-         $('#btn-save').prop("disabled", true);
-       }
- });
-
-
