--- resources/views/categories/index.blade.php
@@ -1,290 +1,0 @@
-@extends('layouts.app')
-
-@section('style')
-<link rel="stylesheet" type="text/css" href="https://cdn.datatables.net/1.10.20/css/dataTables.bootstrap4.min.css">
-@endsection
-
-@section('head_js')
-@endsection
-
-@section('content')
-<div class="container">
-    <div class="row justify-content-center">
-        <div class="col-md-12">
-            <div class="card">
-                <div class="card-header">
-                    {{ __('Category Listing') }}
-                    <button style="float: right; font-weight: 900;" class="btn btn-info btn-sm" type="button"  data-toggle="modal" data-target="#CreateProductModal">
-                        Create Category
-                    </button>
-                </div>
-
-                <div class="card-body">
-                    <div class="table-responsive">
-                        <table class="table table-bordered" id="datatable">
-                            <thead>
-                                <tr>
-                                    <th>Id</th>
-                                    <th>Name</th>
-                                    <th>Slug</th>
-                                    <th width="150" class="text-center">Action</th>
-                                </tr>
-                            </thead>
-                        </table>
-                    </div>
-                </div>
-            </div>
-        </div>
-    </div>
-</div>
-
-<!-- Create Product Modal -->
-<div class="modal" id="CreateProductModal">
-    <div class="modal-dialog">
-        <div class="modal-content">
-            <!-- Modal Header -->
-            <div class="modal-header">
-                <h4 class="modal-title">Product Create</h4>
-                <button type="button" class="close" data-dismiss="modal">&times;</button>
-            </div>
-            <!-- Modal body -->
-            <form>
-            <div class="modal-body">
-                <div class="alert alert-danger alert-dismissible fade show" role="alert" style="display: none;">
-                    <button type="button" class="close" data-dismiss="alert" aria-label="Close">
-                        <span aria-hidden="true">&times;</span>
-                    </button>
-                </div>
-                
-                <div class="alert alert-success alert-dismissible fade show" role="alert" style="display: none;">
-                    <strong>Success!</strong>Product was added successfully.
-                    <button type="button" class="close" data-dismiss="alert" aria-label="Close">
-                        <span aria-hidden="true">&times;</span>
-                    </button>
-                </div>
-                
-                <div class="form-group">
-                    <label for="Name">Name:</label>
-                    <input type="text" class="form-control" name="name" id="name">
-                </div>
-                
-            </div>
-            <!-- Modal footer -->
-            <div class="modal-footer">
-                <button type="button" class="btn btn-success" id="SubmitCreateProductForm">Create</button>
-                <button type="button" class="btn btn-danger" data-dismiss="modal">Close</button>
-            </div>
-            </form>
-        </div>
-    </div>
-</div>
-
-<!-- Edit Product Modal -->
-<div class="modal" id="EditProductModal">
-    <div class="modal-dialog">
-        <div class="modal-content">
-            <!-- Modal Header -->
-            <div class="modal-header">
-                <h4 class="modal-title">Product Edit</h4>
-                <button type="button" class="close modelClose" data-dismiss="modal">&times;</button>
-            </div>
-            <!-- Modal body -->
-            <form>
-            <div class="modal-body">
-                <div class="alert alert-danger alert-dismissible fade show" role="alert" style="display: none;">
-                    <button type="button" class="close" data-dismiss="alert" aria-label="Close">
-                        <span aria-hidden="true">&times;</span>
-                    </button>
-                </div>
-                <div class="alert alert-success alert-dismissible fade show" role="alert" style="display: none;">
-                    <strong>Success!</strong>Product was added successfully.
-                    <button type="button" class="close" data-dismiss="alert" aria-label="Close">
-                        <span aria-hidden="true">&times;</span>
-                    </button>
-                </div>
-                
-                <div id="EditProductModalBody">
-                    
-                </div>
-                
-            </div>
-            <!-- Modal footer -->
-            <div class="modal-footer">
-                <button type="button" class="btn btn-success" id="SubmitEditProductForm">Update</button>
-                <button type="button" class="btn btn-danger modelClose" data-dismiss="modal">Close</button>
-            </div>
-        </form>
-        </div>
-    </div>
-</div>
-
-<!-- Delete Product Modal -->
-<div class="modal" id="DeleteProductModal">
-    <div class="modal-dialog">
-        <div class="modal-content">
-            <!-- Modal Header -->
-            <div class="modal-header">
-                <h4 class="modal-title">Product Delete</h4>
-                <button type="button" class="close" data-dismiss="modal">&times;</button>
-            </div>
-            <!-- Modal body -->
-            <div class="modal-body">
-                <h4>Are you sure want to delete this product?</h4>
-            </div>
-            <!-- Modal footer -->
-            <div class="modal-footer">
-                <button type="button" class="btn btn-danger" id="SubmitDeleteProductForm">Yes</button>
-                <button type="button" class="btn btn-default" data-dismiss="modal">No</button>
-            </div>
-        </div>
-    </div>
-</div>
-@endsection
-
-@push('bot_js')
-<script src="https://code.jquery.com/jquery-3.3.1.js"></script>
-<script src="https://cdnjs.cloudflare.com/ajax/libs/popper.js/1.14.7/umd/popper.min.js"></script>
-<script src="https://maxcdn.bootstrapcdn.com/bootstrap/4.3.1/js/bootstrap.min.js"></script>
-
-<script defer type="text/javascript" src="https://cdn.datatables.net/1.10.20/js/jquery.dataTables.min.js"></script>
-<script defer type="text/javascript" src="https://cdn.datatables.net/1.10.20/js/dataTables.bootstrap4.min.js"></script>
-<script type="text/javascript">
-   $(document).ready(function() {
-        // init datatable.
-        var dataTable = $('#datatable').DataTable({
-            processing: true,
-            serverSide: true,
-            autoWidth: false,
-            pageLength: 5,
-            // scrollX: true,
-            "order": [[ 0, "desc" ]],
-            ajax: "{!! route('category.ajaxIndex') !!}",
-            columns: [
-                {data: 'id', name: 'id'},
-                {data: 'name', name: 'name'},
-                {data: 'slug', name: 'slug'},
-                {data: 'Actions', name: 'Actions',orderable:false,serachable:false,sClass:'text-center'},
-            ]
-        });
-
-        // Create product Ajax request.
-        $('#SubmitCreateProductForm').click(function(e) {
-            e.preventDefault();
-            $.ajaxSetup({
-                headers: {
-                    'X-CSRF-TOKEN': $('meta[name="csrf-token"]').attr('content')
-                }
-            });
-            $.ajax({
-                url: "{!! route('category.ajaxStore') !!}",
-                method: 'post',
-                data: {
-                    name: $('#name').val(),
-                },
-                success: function(result) {
-                    if(result.errors) {
-                        $('.alert-danger').html('');
-                        $.each(result.errors, function(key, value) {
-                            $('.alert-danger').show();
-                            $('.alert-danger').append('<strong><li>'+value+'</li></strong>');
-                        });
-                    } else {
-                        $('.alert-danger').hide();
-                        $('.alert-success').show();
-                        $('#datatable').DataTable().ajax.reload();
-                        setInterval(function(){ 
-                            $('.alert-success').hide();
-                            $('#CreateProductModal').modal('hide');
-                            location.reload();
-                        }, 2000);
-                    }
-                }
-            });
-        });
-
-        // Get single product in EditModel
-        $('.modelClose').on('click', function(){
-            $('#EditProductModal').hide();
-        });
-        var id;
-        $('body').on('click', '#getEditProductData', function(e) {
-            e.preventDefault();
-            $('.alert-danger').html('');
-            $('.alert-danger').hide();
-            id = $(this).data('id');
-            $.ajax({
-                url: "categories/"+id+"/edit",
-                method: 'GET',
-                // data: {
-                //     id: id,
-                // },
-                success: function(result) {
-                    console.log(result);
-                    $('#EditProductModalBody').html(result.html);
-                    $('#EditProductModal').show();
-                }
-            });
-        });
-
-        // Update product Ajax request.
-        $('#SubmitEditProductForm').click(function(e) {
-            e.preventDefault();
-            $.ajaxSetup({
-                headers: {
-                    'X-CSRF-TOKEN': $('meta[name="csrf-token"]').attr('content')
-                }
-            });
-            $.ajax({
-                url: "categories/"+id,
-                method: 'PUT',
-                data: {
-                    name: $('#editName').val(),
-                },
-                success: function(result) {
-                    if(result.errors) {
-                        $('.alert-danger').html('');
-                        $.each(result.errors, function(key, value) {
-                            $('.alert-danger').show();
-                            $('.alert-danger').append('<strong><li>'+value+'</li></strong>');
-                        });
-                    } else {
-                        $('.alert-danger').hide();
-                        $('.alert-success').show();
-                        $('#datatable').DataTable().ajax.reload(null, false);
-                        setInterval(function(){ 
-                            $('.alert-success').hide();
-                            $('#EditProductModal').hide();
-                        }, 2000);
-                    }
-                }
-            });
-        });
-
-        // Delete product Ajax request.
-        var deleteID;
-        $('body').on('click', '#getDeleteId', function(){
-            deleteID = $(this).data('id');
-        })
-        $('#SubmitDeleteProductForm').click(function(e) {
-            e.preventDefault();
-            var id = deleteID;
-            $.ajaxSetup({
-                headers: {
-                    'X-CSRF-TOKEN': $('meta[name="csrf-token"]').attr('content')
-                }
-            });
-            $.ajax({
-                url: "categories/"+id,
-                method: 'DELETE',
-                success: function(result) {
-                    $('#datatable').DataTable().ajax.reload(null, false);
-                        setInterval(function(){ 
-                            $('.alert-success').hide();
-                            $('#DeleteProductModal').hide();
-                        }, 1000);
-                }
-            });
-        });
-    });
-</script>
-@endpush
