+++ routes/web.php
@@ -19,19 +19,8 @@ Auth::routes();
 
 Route::get('/home', 'HomeController@index')->name('home');
 
-Auth::routes();
-
-Route::get('/home', 'HomeController@index')->name('home');
-
-// Category testing
+// ADMIN - Category CRUD
 Route::resource('/category', 'CategoryController');
-Route::get('categories', 'CategoryController@ajaxIndex')->name('category.ajaxIndex');
-Route::post('categories', 'CategoryController@ajaxStore')->name('category.ajaxStore');
-
-Route::get('categories/{category}/edit', 'CategoryController@ajaxEdit')->name('category.ajaxEdit');
-Route::put('categories/{category}', 'CategoryController@ajaxUpdate')->name('category.ajaxUpdate');
-
-Route::delete('categories/{category}', 'CategoryController@ajaxDelete')->name('category.ajaxDelete');
-Route::delete('categories/{category}/hardDelete', 'CategoryController@ajaxHardDelete')->name('category.ajaxHardDelete');
-Route::patch('categories/{category}/restoreDelete', 'CategoryController@ajaxRestoreDelete')->name('category.ajaxRestoreDelete');
-Route::get('categories/trash', 'CategoryController@getTrashRecords')->name('category.getTrashRecords');
+Route::delete('/category/{category}/emptyTrash', 'CategoryController@emptyTrash')->name('category.emptyTrash');
+Route::patch('/category/{category}/restoreTrash', 'CategoryController@restoreTrash')->name('category.restoreTrash');
+Route::get('/category/trash/sd', 'CategoryController@getTrashRecords')->name('category.trash');
